+++ src/components/forms/ConfigureSourceForm/ConfigureSourceForm.tsx
@@ -3,8 +3,7 @@ import { Formik } from "formik";
 import { SingleSelectOption } from "@instill-ai/design-system";
 
 import { FormBase, SingleSelect } from "@/components/formik";
-import { ConnectorIcon } from "@/components/ui";
-import { PrimaryButton } from "@/components/ui/Buttons";
+import { ConnectorIcon, PrimaryButton } from "@/components/ui";
 import { SourceWithPipelines } from "@/lib/instill";
 import { Nullable } from "@/types/general";
