+++ src/components/forms/CreatePipelineForm/SetupDestinationStep/CreateNewDestinationFlow.tsx
@@ -2,7 +2,6 @@ import { FC, useState, useEffect, useMemo } from "react";
 import { useFormikContext } from "formik";
 import { SingleSelectOption } from "@instill-ai/design-system";
 
-import { PrimaryButton } from "@/components/ui/Buttons";
 import { SingleSelect, TextField } from "../../../formik";
 import {
   StepNumberState,
@@ -12,7 +11,7 @@ import {
   useCreateDestination,
   useDestinationDefinitions,
 } from "@/services/connector";
-import { ConnectorIcon } from "@/components/ui";
+import { ConnectorIcon, PrimaryButton } from "@/components/ui";
 import { CreateDestinationPayload } from "@/lib/instill";
 
 export type CreateNewDestinationFlowProps = StepNumberState;
