+++ src/components/forms/CreatePipelineForm/SetupModelStep/CreateNewModelFlow.tsx
@@ -20,7 +20,6 @@ import {
   TextField,
   UploadFileField,
 } from "../../../formik";
-import { PrimaryButton } from "@/components/ui/Buttons";
 import {
   StepNumberState,
   CreatePipelineFormValues,
@@ -32,7 +31,7 @@ import {
   useModelDefinitions,
   useModelInstances,
 } from "@/services/model";
-import { ModelDefinitionIcon } from "@/components/ui";
+import { ModelDefinitionIcon, PrimaryButton } from "@/components/ui";
 import {
   CreateGithubModelPayload,
   CreateLocalModelPayload,
