+++ src/components/forms/CreatePipelineForm/SetupModelStep/UseExistingModelFlow.tsx
@@ -10,7 +10,7 @@ import {
 import { useFormikContext } from "formik";
 import { SingleSelectOption } from "@instill-ai/design-system";
 
-import { PrimaryButton } from "@/components/ui/Buttons";
+import { PrimaryButton } from "@/components/ui";
 import { SingleSelect } from "../../../formik";
 import {
   StepNumberState,
