+++ src/components/ui/index.ts
@@ -61,3 +61,5 @@ export * from "./TableCells";
 export * from "./TablePlaceholders";
 
 export * from "./Tables";
+
+export * from "./Buttons";
